Add files via upload
The code has been made more modular. The code for creating and manipulating a network of vertices, edges, faces, facets, etc has been moved to geometry.c. 

The geo object was created which encapsulates geometry data and would serve as the main object for someone intending to use this codebase as a library. 

Drawing of geo objects with openGL was placed in geodraw.c

A makefile was added to connect the pieces.

The main file, polytope.c, now mainly contains the high-level control which sets up the visualization. Additionally, a menu was added which is accessible by right-clicking. The menu contains many basic controls along with a list of pre-loaded schlafli symbols.